\name{alr4-package}
\alias{alr4-package}
\alias{alr4}
\docType{package}
\title{
Applied Linear Regression 4th Edition Data Sets
}
\description{
This package accompanies S. Weisberg, Applied Regression Regression,
Fourth Edition, Wiley. 2014.  Associated functions have been moved to the
\code{\link{car}} and \code{\link{effects}} packages.
}
\details{
\tabular{ll}{
Package: \tab alr4\cr
Version: \tab 1.0.0\cr
Date: \tab 2013/01/03\cr
Depends: \tab R (>= 2.1.1), car, effects\cr
License: \tab GPL (>= 2)\cr
URL: \tab http://www.r-project.org, http://www.stat.umn.edu/alr\cr
}
} 


\author{
Sanford Weisberg <sandy@umn.edu> 
} 
\keyword{ package }
