\name{baeskel}
\alias{baeskel}
\docType{data}
\title{Surface tension}
\description{The data in the file were collected in a study of the effect of dissolved sulfur on the surface tension of liquid copper (Baes and Kellogg, 1953)}
\format{This data frame contains the following columns:
 \describe{
    \item{Sulfur}{
Weight percent sulfur 
    }
    \item{Tension}{
Decrease in surface tension, dynes/cm 
    }
 }
}

\source{
Baes, C. and Kellogg,  H. (1953).  Effect of dissolved sulphur on the
surface tension of liquid copper.  \emph{J. Metals}, 5, 643-648.}

\references{Weisberg, S. (2014).  \emph{Applied Linear Regression}, 4th
edition. Hoboken NJ: Wiley.}

\examples{
head(baeskel)
}
\keyword{datasets}
