\name{brains}
\alias{brains}
\docType{data}
\title{Mammal brain weights}
\description{
The data provided gives the average body weight in kilograms and the average brain weight in grams for sixty-two species of mammals.}

\format{This data frame uses species names as row labels and
contains the following columns:
 \describe{
    \item{BrainWt}{
Brain weight, grams 
    }
    \item{BodyWt}{
Body weight, kg 
    }
 }
}

\source{Allison, T. and Cicchetti, D. (1976). Sleep in mammals: Ecology and constitutional correlates. \emph{Science}, 194, 732-734.}
\references{Weisberg, S. (2014). \emph{Applied Linear Regression}, 4th
edition.  Hoboken NJ: Wiley.}
\examples{
head(brains)
}
\keyword{datasets}
