\name{cathedral}
\alias{cathedral}
\docType{data}
\title{Cathedrals}
\description{
Heights and lengths of Gothic and Romanesque cathedrals.
}
\format{This data frame uses cathedral names as row label 
andcontains the following columns:
 \describe{
    \item{Type}{
Romanesque or Gothic
    }
    \item{Height}{
Total height, feet 
    }
    \item{Length}{
Total length, feet 
    }
 }
}

\source{Stephen Jay Gould}
\references{Weisberg, S. (2014). \emph{Applied Linear Regression}, 4th
edition. Hoboken NJ: Wiley.}
\examples{
head(cathedral)
}
\keyword{datasets}
