\name{dwaste}
\alias{dwaste}
\docType{data}
\title{Crock data.}
\description{
An experiment was conducted to study the \emph{O2UP}, oxygen uptake in milligrams of oxygen per minute, given five chemical measurements: biological oxygen demand (BOD), total Kjeldah nitrogen (TKN), total solids (TS), total vital solids (TVS), which is a component of TS, and chemical oxygen demand (COD), each measured in milligrams per liter (Moore, 1975).
}
\format{This data frame contains the following columns:
 \describe{
    \item{Day}{Day number
 
    }
    \item{BOD}{Biological oxygen demand
 
    }
    \item{TKN}{Total Kjeldahl nitrogen
 
    }
    \item{TS}{Total Solids
 
    }
    \item{TVS}{
    Total volatile solids  
 
    }
    \item{COD}{Chemical oxygen demand
 
    }
    \item{O2UP}{Oxygen uptake
 
    }
 }
}

\source{Moore, J. (1975). Total Biomedical Oxygen Demand of Animal Manures. Unpublished Ph. D. disseration, University of Minnesota.}
\references{Weisberg, S. (2014). \emph{Applied Linear Regression}, 4th
edition. Hoboken NJ: Wiley.}
\examples{
head(dwaste)
}
\keyword{datasets}
