\name{mantel}
\alias{mantel}
\docType{data}
\title{ Mantel's artifical data for stepwise regression}
\description{
  An artificial data set suggested by N. Mantel to illustrate stepwise
  regression methods.
}
\format{
  A data frame with 5 observations on the following 4 variables.
  \describe{
    \item{Y}{the response}
    \item{X1}{predictor 1}
    \item{X2}{predictor 2}
    \item{X3}{predictor 3}
  }
}
\source{
 Mantel, N. (1970).
Why stepdown procedures in variable selection? 
\emph{Technometrics}, 12, 621--625.
}
\references{Weisberg, S. (2014). \emph{Applied Linear Regression}, 4th
edition. Hoboken NJ: Wiley.}
\examples{
head(mantel)
}
\keyword{datasets}
