\name{Mitchell}
\alias{Mitchell}
\docType{data}
\title{Mitchell soil temperature}
\description{
Data collected by Kenneth G. Hubbard on soil temperature at 20 cm depth in
Mitchell, Nebraska for 17 years (1976-1992)
The variable month is the month number.
}
\format{This data frame contains the following columns:
 \describe{
    \item{Month}{
 Months beginning Jan, 1976
    }
    \item{Temp}{
 Average soil temperature, degrees C
    }
 }
}

\source{Kenneth G. Hubbard}
\references{Weisberg, S. (2014). \emph{Applied Linear Regression}, 4th
edition. Hoboken NJ: Wiley.}
\examples{
head(Mitchell)
}
\keyword{datasets}
