\name{shocks}
\alias{shocks}
\docType{data}
\title{Small electric shocks in dairy cows}
\description{
Results of a small experiment to learn about the effects of small electric shocks on dairy cows.
}
\format{
  A data frame with 6 observations on the following 3 variables.
  \describe{
    \item{Intensity}{Shock level, milliamps}
    \item{m}{Number of trials}
    \item{Y}{Number of times a positive reaction was observed}
  }
}
\source{
 R. Norell
}
\references{Weisberg, S. (2014). \emph{Applied Linear Regression}, 4th
edition. Hoboken NJ: Wiley.}
\examples{
head(shocks)
}
\keyword{datasets}
