\name{snake}
\alias{snake}
\docType{data}
\title{Snake river data}
\description{
The data give the water content of snow and the water yield in inches in the Snake River watershed in Wyoming.
}
\format{This data frame contains the following columns:
 \describe{
    \item{X}{
 water content of snow
    }
    \item{Y}{
 water yield from April to July
    }
 }
}

\source{Wilm, H. G. (1950). Statistical control in hydrologic forecasting.
\dQuote{Res. Notes}, 61, Pacific Northwest Forest Range Experiment Station,
Oregon.}

\references{Weisberg, S. (2014). \emph{Applied Linear Regression}, 4th
edition. Hoboken NJ: Wiley.}
\examples{
head(snake)
}
\keyword{datasets}
