\name{snowgeese}
\alias{snowgeese}
\docType{data}
\title{Snow geese}
\description{
Counts of flocks of snow geese.
}
\format{This data frame contains the following columns:
 \describe{
    \item{photo}{
 Photo count
    }
    \item{obs1}{
 Observer 1 count (aerial)
    }
    \item{obs2}{
 Observer 2 count (aerial)
    }
 }
}

\source{Dennis Cook}
\references{Weisberg, S. (in press). \emph{Applied Linear Regression}, 4th
edition. Hoboken NJ: Wiley.}
\examples{
head(snowgeese)
}
\keyword{datasets}
