\name{swan96}
\alias{swan96}
\docType{data}
\title{Black crappie study on Swan Lake, Minnesota}
\description{
Log catch per unit effect of 200 mm or longer black crappies was recored 27 times over the course
of 1996 on Swan Lake, Minnesota.
}
\format{
  A data frame with 27 observations on the following 2 variables.
  \describe{
    \item{Day}{Number of days after June 16, 1996}
    \item{LCPUE}{log of the catch of 200 mm or longer black crappies per unit effort (WHAT IS THE BASE?)}
  }
}

\source{
Minnesota Department of Natural Resources
}
\references{Weisberg, S. (2014). \emph{Applied Linear Regression}, 4th
edition. Hoboken NJ: Wiley.}
\examples{
head(swan96)
}
\keyword{datasets}
