\name{twins}
\alias{twins}
\docType{data}
\title{Burt's twin data}
\description{
The given data are IQ scores from identical twins; one raised in a foster home, and the other raised by birth parents.
}
\format{This data frame contains the following columns:
 \describe{
    \item{C}{
 Social class, C1=high, C2=medium, C3=low, a factor
    }
    \item{IQb}{
 biological
    }
    \item{IQf}{
 foster
    }
 }
}

\source{Burt, C. (1966). The genetic estimation of differences in intelligence: A study of monozygotic twins reared together and apart. \emph{Br. J. Psych.}, 57, 147-153.}
\references{Weisberg, S. (2014). \emph{Applied Linear Regression}, 4th
edition. Hoboken NJ: Wiley.}
\examples{
head(twins)
}
\keyword{datasets}
