\name{walleye}
\alias{walleye}
\docType{data}
\title{Walleye length at age}
\description{
These data give length and age for over 3000 walleye (a type of fish) captured in Butternut
Lake, Wisconsin, in three periods with different management method in place.
}
\format{
  A data frame with 3198 observations on the following 3 variables.
  \describe{
    \item{age}{Age of the fish, years}
    \item{length}{Length, mm}
    \item{period}{1 = pre 1990, 2 = 1991-1996, 3=1997-2000}
  }
}
\source{
Michelle LeBeau
}
\references{Weisberg, S. (2014). \emph{Applied Linear Regression}, 4th edition. Hoboken NJ: Wiley.}
\examples{
head(walleye)
}
\keyword{datasets}
