\name{wm3}
\alias{wm3}
\docType{data}
\title{Binned wind speed data}
\description{
For the windspeed data in the file \code{wm2}, this is the number of observations in each of the 
bins, and the average windspeed in that bin
}
\format{
  A data frame with 16 observations on the following 2 variables.
  \describe{
    \item{bin.count}{Number of observations in the bin}
    \item{RSpd}{Average windspeed in the bin}
  }
}
\references{Weisberg, S. (in press). \emph{Applied Linear Regression}, 4th edition. Hoboken NJ: Wiley.
}
\examples{
head(wm3)
}
\keyword{datasets}
