\name{Htwt}
\alias{Htwt}
\docType{data}
\title{Artificial height and weight data}
\description{
The data for this table are a sample size of ten 18-year old girls taken from the study that was conducted by Tuddenham and Snyder (1954).
}
\format{This data frame contains the following columns:
 \describe{
    \item{ht}{
 Height (cm) at age 18
    }
    \item{wt}{
 Weight (kg) at age 18
    }
 }
}

\source{Tuddenham, R., and Snyder, M. (1954). Physical growth of California boys and girls from birth to age 18. \emph{California Publications on Child Development}, 1, 183-364.}
\references{Weisberg, S. (2014). \emph{Applied Linear Regression}, 4th
edition. Hoboken NJ: Wiley.}
\examples{
head(Htwt)
}
\keyword{datasets}
